requirements.txt
Django==4.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3